requirements.txt
Flask
PyYAML
Pillow
